services/inference-worker/requirements.txt
redis==5.0.8
prometheus-client==0.20.0

opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation-redis==0.47b0